.asciiv v:2 — fixed-fps timeline + WYSIWYG per-cell colour (sync redesign)
Root cause of the chronic embed A/V drift: v:1 shipped raw uneven capture
timestamps (times[]) and made the player reconcile THREE clocks — the frame
origin (recStart, stamped after mr.start() returned, i.e. later than the
audio's real t=0: the "frames run ahead" bug), the audio recorder's own start,
and three period estimates (audioEl.duration / durationMs / the times span).
Every mismatch surfaced as a new sync bug (freeze, missing beginning, ahead).

v:2 removes the mapping instead of patching it:
- recStart is stamped at mr.onstart (the audio recording's actual t=0).
- The bake resamples the uneven rVFC capture onto a uniform fps grid aligned
  to that origin (resampleToFps: each grid instant takes the frame that was on
  screen then; frame 0 before the first capture, so the seek-to-0 start and
  the no-gaps property are inherent). durationMs := frameCount/fps*1000.
- Playback: frame = floor(phase*fps) clamped, wrapping on the audio's own
  duration — one clock, one rate, bounded non-accumulating error, no times[].
- Cells are packed u16 (char-index | colourKey<<4) capturing what the live
  renderer displayed — gray level, or the 125-cube index when saturation>0,
  so saturation now carries into embeds (WYSIWYG; buildRows2 reproduces the
  live markup byte-for-byte) and is back in embedSig. Still a tiny fixed
  class palette: run-merging perf untouched.

v:1 files on S3 keep decoding exactly as before (decoder dispatches on
header.v; frameAt/even-phase playback paths retained). validHeader fails
closed on unknown versions and malformed cube; the gzip-bomb cap covers the
u16 stream. New tests/asciiv2.test.ts proves the resample against ground
truth, the bounded-sync-error composition, v:2 round-trip byte-stability,
the playback mapping (phase-locked over 100k loops), and v:1 compatibility.

+++ src/state.ts
@@ -102,7 +102,7 @@ export interface Rt {
   computing: boolean;
   firstPaintPending: boolean;
   recording: boolean;
-  recFrames: Uint8Array[];
+  recFrames: Uint16Array[]; // packed .asciiv v:2 cells (char-index | colourKey<<4), one grid per captured frame
   recTimes: number[];
   recStart: number;
   baking: boolean;
